+++ src/Phan/Plugin/Internal/CallableParamPlugin.php
@@ -8,6 +8,7 @@ use Phan\AST\UnionTypeVisitor;
 use Phan\CodeBase;
 use Phan\Config;
 use Phan\Language\Context;
+use Phan\Language\Element\Func;
 use Phan\Language\Element\FunctionInterface;
 use Phan\Language\Type;
 use Phan\Language\Type\CallableInterface;
@@ -15,6 +16,8 @@ use Phan\Language\Type\ClassStringType;
 use Phan\Plugin\ConfigPluginSet;
 use Phan\PluginV2;
 use Phan\PluginV2\AnalyzeFunctionCallCapability;
+use Phan\PluginV2\HandleLazyLoadInternalFunctionCapability;
+use function count;
 
 /**
  * NOTE: This is automatically loaded by phan. Do not include it in a config.
@@ -23,7 +26,8 @@ use Phan\PluginV2\AnalyzeFunctionCallCapability;
  * This would be difficult.
  */
 final class CallableParamPlugin extends PluginV2 implements
-    AnalyzeFunctionCallCapability
+    AnalyzeFunctionCallCapability,
+    HandleLazyLoadInternalFunctionCapability
 {
 
     /**
@@ -53,7 +57,7 @@ final class CallableParamPlugin extends PluginV2 implements
                 // Fetch possible functions. As an intentional side effect, this warns about invalid callables.
                 // TODO: Check if the signature allows non-array callables? Not sure of desired semantics.
                 $function_like_list = UnionTypeVisitor::functionLikeListFromNodeAndContext($code_base, $context, $arg, true);
-                if (\count($function_like_list) === 0) {
+                if (count($function_like_list) === 0) {
                     // Nothing to do
                     continue;
                 }
@@ -73,7 +77,7 @@ final class CallableParamPlugin extends PluginV2 implements
 
                 // Fetch possible classes. As an intentional side effect, this warns about invalid/undefined class names.
                 $class_list = UnionTypeVisitor::classListFromClassNameNode($code_base, $context, $arg);
-                if (\count($class_list) === 0) {
+                if (count($class_list) === 0) {
                     // Nothing to do
                     continue;
                 }
@@ -91,6 +95,37 @@ final class CallableParamPlugin extends PluginV2 implements
     }
 
     /**
+     * @return ?Closure(CodeBase,Context,FunctionInterface,array):void
+     */
+    private static function generateClosureForFunctionInterface(FunctionInterface $function)
+    {
+        $callable_params = [];
+        $class_params = [];
+        foreach ($function->getParameterList() as $i => $param) {
+            // If there's a type such as Closure|string|int, don't automatically assume that any string or array passed in is meant to be a callable.
+            // Explicitly require at least one type to be `callable`
+            if ($param->getUnionType()->hasTypeMatchingCallback(static function (Type $type) : bool {
+                // TODO: More specific closure for CallableDeclarationType
+                return $type instanceof CallableInterface;
+            })) {
+                $callable_params[] = $i;
+            }
+            if ($param->getUnionType()->hasTypeMatchingCallback(static function (Type $type) : bool {
+                return $type instanceof ClassStringType;
+            })) {
+                $class_params[] = $i;
+            }
+        }
+
+        if (count($callable_params) === 0 && count($class_params) === 0) {
+            return null;
+        }
+        // Generate a de-duplicated closure.
+        // fqsen can be global_function or ClassName::method
+        return self::generateClosure($callable_params, $class_params);
+    }
+
+    /**
      * @return array<string,\Closure>
      * @phan-return array<string,Closure(CodeBase,Context,FunctionInterface,array):void>
      */
@@ -98,29 +133,12 @@ final class CallableParamPlugin extends PluginV2 implements
     {
         $result = [];
         $add_callable_checker_closure = static function (FunctionInterface $function) use (&$result) {
-            $callable_params = [];
-            $class_params = [];
-            foreach ($function->getParameterList() as $i => $param) {
-                // If there's a type such as Closure|string|int, don't automatically assume that any string or array passed in is meant to be a callable.
-                // Explicitly require at least one type to be `callable`
-                if ($param->getUnionType()->hasTypeMatchingCallback(static function (Type $type) : bool {
-                    // TODO: More specific closure for CallableDeclarationType
-                    return $type instanceof CallableInterface;
-                })) {
-                    $callable_params[] = $i;
-                }
-                if ($param->getUnionType()->hasTypeMatchingCallback(static function (Type $type) : bool {
-                    return $type instanceof ClassStringType;
-                })) {
-                    $class_params[] = $i;
-                }
-            }
-            if (\count($callable_params) === 0 && \count($class_params) === 0) {
-                return;
-            }
             // Generate a de-duplicated closure.
             // fqsen can be global_function or ClassName::method
-            $result[$function->getFQSEN()->__toString()] = self::generateClosure($callable_params, $class_params);
+            $closure = self::generateClosureForFunctionInterface($function);
+            if ($closure) {
+                $result[$function->getFQSEN()->__toString()] = $closure;
+            }
         };
 
         $add_another_closure = static function (string $fqsen, Closure $closure) use (&$result) {
@@ -167,7 +185,22 @@ final class CallableParamPlugin extends PluginV2 implements
     }
 
     /**
-     * @phan-return array<string,Closure(CodeBase,Context,FunctionInterface,array):void>
+     * When a function is loaded into the CodeBase for the first time during analysis
+     * (e.g. `register_shutdown_function()`, this is called to conditionally add any checkers for callable/closure.
+     */
+    public function handleLazyLoadInternalFunction(
+        CodeBase $unused_code_base,
+        Func $function
+    ) {
+        $closure = self::generateClosureForFunctionInterface($function);
+        if ($closure) {
+            $function->addFunctionCallAnalyzer($closure);
+        }
+    }
+
+    /**
+     * @return array<string,Closure(CodeBase,Context,FunctionInterface,array):void>
+     * @override
      */
     public function getAnalyzeFunctionCallClosures(CodeBase $code_base) : array
     {