+++ tests/plugin_test/src/094_shutdown_function.php
@@ -1,0 +1,4 @@
+<?php
+
+register_shutdown_function('notafunction');  // Fails to emit PhanUndeclaredFunctionInCallable
+register_shutdown_function(['MissingClass', 'notafunction']);  // Fails to emit PhanUndeclaredFunctionInCallable or anything else